Renamed DX11 versions of standalone modules
